stylebend/requirements.txt
streamlit==1.36.0
numpy>=1.24,<3
Pillow>=10.0.0,<11
opencv-python>=4.9.0
scikit-image>=0.22
click>=8.1
tqdm>=4.66

# Torch stack: pinned but install via CUDA index-url for GPU wheels
# See commands below to install CUDA 12.1 or 11.8 builds.
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

# Perceptual similarity for projection
lpips==0.1.4 